Integrate swipe page and change tab icons
